feat: tweak interface to signal module
* Capability to pass model kwargs from config file
to model class constructor.

* Remove SimulationTemplateQSONGC class.
The pattern used to look up the simulation
templates is now provided in the config file
under model_kwargs.

* Remove signal.stack_av, signal.stack_av_group,
and signal.load_all and replace with the
equivalent functions from utils, which allows
selecting the polarisations and sorting of the
frequency axis in the same manner as the data and mocks.

* Update default boundaries for SimulationTemplate class.

* Create utils.load_pol function that load a data file into
the proper container with the requested polarisations.

* Create utils.average_stacks to average over mock catalogs.
Use that when constructing average template.

* Add fixed dataset that indicates the parameters held fixed
during the fit.

* Fixed a bug and can now hold parameters fixed by setting
the fixed keyword to True in the param_spec.

* Combine the _param_spec from all subclasses.