feat(chisq_test): code that was used for stacking analysis
This includes all code that was used for the stacking analysis
but not commited to the repository.  This branches off the
jrs/fogc branch.  In particular the module chisq_test implements
a statistical test that was used to test consistency between
polarisations and even-odd days.

The original commit from the ss/stacking branch was manually
modified by Simon to resolve conflicts with newer commits from
the jrs/fogc and auto branches.