feat: add force_real option
This commit adds a force_real argument to several functions,
which takes the real part of input datasets and associated
metadata. This is designed for cases where power spectra are
stored as complex numbers, but should have either zero or
negligible imaginary part. Explicitly taking the real part
avoids complex-to-real casting warnings and unintended
consequences of having mostly-real quantities stored as complex.