fix(get_powerspectrum1d_LOO_invcovariance): remove pol selection
Selecting polarizations for the covariance is unneeded here, since
the required polarizations were already selected in the input mocks
array. This actually caused problems if only using QQ, since the
code attempted to select the index-1 pol from an array with a pol
axis with length 1 (because QQ would already have been subselected).